OrderBy refactoring
