+++ src/codemagic/google/resource_managers/mixins/acting_manager_mixin.py
@@ -5,7 +5,6 @@ from abc import ABC
 from abc import abstractmethod
 from typing import TYPE_CHECKING
 from typing import Any
-from typing import ClassVar
 from typing import Dict
 from typing import Generic
 from typing import Type
@@ -15,9 +14,9 @@ from googleapiclient import errors
 from googleapiclient.http import HttpRequest
 from oauth2client.client import Error as OAuth2ClientError
 
-from ...errors import GoogleApiHttpError
-from ...errors import GoogleAuthenticationError
-from ...errors import GoogleClientError
+from codemagic.google.errors import GoogleApiHttpError
+from codemagic.google.errors import GoogleAuthenticationError
+from codemagic.google.errors import GoogleClientError
 
 if TYPE_CHECKING:
     from ...resources.resource import Resource
@@ -26,28 +25,28 @@ ResourceT = TypeVar('ResourceT', bound='Resource')
 
 
 class ActingManagerMixin(Generic[ResourceT], ABC):
-    manager_action: ClassVar[str]
+    _logger: logging.Logger
+
+    @property
+    @abstractmethod
+    def manager_action(self):
+        ...
 
     @property
     @abstractmethod
     def resource_type(self) -> Type[ResourceT]:
         ...
 
-    @property
-    @abstractmethod
-    def logger(self) -> logging.Logger:
-        ...
-
     def _execute_request(self, request: HttpRequest) -> Dict[str, Any]:
         try:
             return request.execute()
         except OAuth2ClientError as e:
-            self.logger.exception(f'Failed to {self.manager_action} Firebase {self.resource_type.get_label()}')
+            self._logger.exception(f'Failed to {self.manager_action} Firebase {self.resource_type.get_label()}')
             raise GoogleAuthenticationError(str(e)) from e
         except errors.HttpError as e:
-            self.logger.exception(f'Failed to {self.manager_action} Firebase {self.resource_type.get_label()}')
+            self._logger.exception(f'Failed to {self.manager_action} Firebase {self.resource_type.get_label()}')
             reason = e.reason  # type: ignore
             raise GoogleApiHttpError(reason) from e
         except errors.Error as e:
-            self.logger.exception(f'Failed to {self.manager_action} Firebase {self.resource_type.get_label()}')
+            self._logger.exception(f'Failed to {self.manager_action} Firebase {self.resource_type.get_label()}')
             raise GoogleClientError(str(e)) from e
