+++ src/codemagic/google/resource_managers/resource_manager.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import logging
 from abc import ABC
 from abc import abstractmethod
 from typing import TYPE_CHECKING
@@ -8,7 +7,6 @@ from typing import Generic
 from typing import Type
 from typing import TypeVar
 
-from codemagic.models.enums import ResourceEnum
 from codemagic.utilities import log
 
 if TYPE_CHECKING:
@@ -20,10 +18,6 @@ ResourceT = TypeVar('ResourceT', bound='Resource')
 
 
 class ResourceManager(Generic[ResourceT], ABC):
-    class OrderBy(ResourceEnum):
-        create_time_desc = 'createTimeDesc'
-        create_time_asc = 'createTime'
-
     def __init__(self, firebase_app_distribution: FirebaseAppDistributionResource):
         self._firebase_app_distribution = firebase_app_distribution
         self._logger = log.get_file_logger(self.__class__)
@@ -32,7 +26,3 @@ class ResourceManager(Generic[ResourceT], ABC):
     @abstractmethod
     def resource_type(self) -> Type[ResourceT]:
         ...
-
-    @property
-    def logger(self) -> logging.Logger:
-        return self._logger
