+++ src/codemagic/google/resources/__init__.py
@@ -1,3 +1,4 @@
 from .release import Release
 from .release_notes import ReleaseNotes
+from .resource import OrderBy
 from .resource import Resource
